♻️ refactor(yttypes): youtube types moved to dedicated src directory
