+++ app/youtubers/[channelId]/page.tsx
@@ -1,9 +1,9 @@
 import { compareDesc } from 'date-fns';
 import { notFound } from 'next/navigation';
-import type { YouTubeVideoItem } from '@/app/actions/fetchYouTube';
 import { getAllChannels, getChannel } from '@/app/actions/handleYTData';
 import Blog from '@/components/Blog/Blog';
 import { siteConfig } from '@/config/site';
+import type { YouTubeVideoItem } from '@/src/types';
 import { absoluteUrl } from '@/src/util/utils';
 import type { ResolvingMetadata } from 'next';
